Add AccountUnitTests and moved Candelete/CanGetAll/CanCreate there instead.
